+++ exporters/zipkin/internal/matchers/expectation.go
@@ -15,7 +15,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package matchers // import "go.opentelemetry.io/otel/internal/matchers"
+package matchers // import "go.opentelemetry.io/otel/exporters/zipkin/internal/matchers"
 
 import (
 	"fmt"